feat(mods): Modularize output formats and organize modules
