+++ src/main/java/com/ignfab/minalac/generator/MinalacGenerator.java
@@ -13,13 +13,10 @@ import com.ignfab.minalac.generator.exceptions.GenerationFailedException;
 import com.ignfab.minalac.generator.exceptions.TransformException;
 import com.ignfab.minalac.generator.generation.Generation;
 import com.ignfab.minalac.generator.generation.GenerationTile;
-import com.ignfab.minalac.generator.outputs.minecraft.MCVoxelWorld;
-import com.ignfab.minalac.generator.outputs.minetest.MTVoxelWorld;
-import com.ignfab.minalac.generator.parameters.OutputFormat;
+import com.ignfab.minalac.generator.modules.minecraft.MinecraftOutputModule;
+import com.ignfab.minalac.generator.modules.minetest.LuantiOutputModule;
 import com.ignfab.minalac.generator.parameters.ParamsParser;
 import com.ignfab.minalac.generator.parameters.ParseException;
-import com.ignfab.minalac.generator.parameters.placeables.voxels.MCVoxelParams;
-import com.ignfab.minalac.generator.parameters.placeables.voxels.MTVoxelParams;
 import com.ignfab.minalac.generator.parameters.processors.FloatMatrixProcessorParams;
 import com.ignfab.minalac.generator.parameters.processors.GeoToolsVectorProcessorParams;
 import com.ignfab.minalac.generator.parameters.processors.OsmProcessorParams;
@@ -102,6 +99,11 @@ public final class MinalacGenerator {
 
         ModulesLoader loader = new ModulesLoader();
 
+        // Built-in modules
+        loader.add(new LuantiOutputModule());
+        loader.add(new MinecraftOutputModule());
+
+        // External modules
         if (cli.modulesPath() != null)
             loader.loadModulesDirectory(cli.modulesPath().toFile());
 
@@ -110,9 +112,6 @@ public final class MinalacGenerator {
         // Generation parsing
         ParamsParser parser = new ParamsParser();
 
-        // Register game formats
-        parser.registerFormat("minecraft", new OutputFormat(() -> new MCVoxelWorld(destination), MCVoxelParams.class, MCVoxelParams::new));
-        parser.registerFormat("minetest", new OutputFormat(() -> new MTVoxelWorld(destination), MTVoxelParams.class, MTVoxelParams::new));
 
         // TODO: Static method that provides a ParamsParser with all default renderers
         // If those name values are modified, update the documentation accordingly
@@ -154,7 +153,7 @@ public final class MinalacGenerator {
 
         modules.registerParams(parser);
 
-        Generation generation = parser.parse(parameters).create(maxTileSize);
+        Generation generation = parser.parse(parameters).create(destination, maxTileSize);
 
         System.out.printf("Generation initialization took %ds.%n", Duration.between(initializationStart, Instant.now()).toSeconds());
         if (cli.generationDisabled()) {
