+++ src/main/java/com/ignfab/minalac/generator/modules/minetest/Block.java
@@ -1,4 +1,4 @@
-package com.ignfab.minalac.generator.outputs.minetest;
+package com.ignfab.minalac.generator.modules.minetest;
 
 import java.util.HashMap;
