+++ src/main/java/com/ignfab/minalac/generator/modules/minetest/MTVoxelParams.java
@@ -1,11 +1,10 @@
-package com.ignfab.minalac.generator.parameters.placeables.voxels;
+package com.ignfab.minalac.generator.modules.minetest;
 
 import java.beans.ConstructorProperties;
 
 import com.fasterxml.jackson.annotation.JsonSetter;
 import com.fasterxml.jackson.annotation.Nulls;
 
-import com.ignfab.minalac.generator.outputs.minetest.MTVoxel;
 import com.ignfab.minalac.generator.parameters.placeables.PlaceableParams;
 import com.ignfab.minalac.generator.placeables.Placeable;
 import com.ignfab.minalac.generator.utils.random.Seed;
