+++ src/test/java/com/ignfab/minalac/generator/modules/minetest/MTVoxelParamsTest.java
@@ -1,4 +1,4 @@
-package com.ignfab.minalac.generator.parameters.placeables.voxels;
+package com.ignfab.minalac.generator.modules.minetest;
 
 import org.junit.jupiter.api.Test;
