+++ src/test/java/com/ignfab/minalac/generator/parameters/OutputFormatTest.java
@@ -19,7 +19,7 @@ public class OutputFormatTest {
     @DisplayName("Simple constructor test")
     public void testConstructor() {
         assertDoesNotThrow(() -> new OutputFormat(
-            () -> null,
+            (destination) -> null,
             TestingVoxelParams.class,
             TestingVoxelParams::new
         ));
@@ -29,7 +29,7 @@ public class OutputFormatTest {
     @DisplayName("Test createVoxelParams using shortcut")
     public void testCreateVoxelParamsShortcut() {
         OutputFormat format = new OutputFormat(
-            () -> null,
+            (destination) -> null,
             TestingVoxelParams.class,
             TestingVoxelParams::new
         );
@@ -43,7 +43,7 @@ public class OutputFormatTest {
     @DisplayName("Test createVoxelParams using default class")
     public void testCreateVoxelParamsDefault() throws JacksonException {
         OutputFormat format = new OutputFormat(
-            () -> null,
+            (destination) -> null,
             TestingVoxelParams.class,
             TestingVoxelParams::new
         );
@@ -63,7 +63,7 @@ public class OutputFormatTest {
     @DisplayName("Test createVoxelParams with null default and shortcut")
     public void testCreateVoxelParamsNull() throws JacksonException {
         OutputFormat format = assertDoesNotThrow(() -> new OutputFormat(
-            () -> null,
+            (destination) -> null,
             null,
             null
         ));
